fix: started addressing feedback
