+++ src/components/RadioGroup/index.ts
@@ -1,4 +1,4 @@
-import { RadioGroup, Radio } from './RadioGroup';
+import RadioGroup from './RadioGroup';
 import * as CONSTANTS from './RadioGroup.constants';
 import { Props, GroupProps } from './RadioGroup.types';
 
@@ -6,5 +6,4 @@ export { CONSTANTS as RADIO_GROUP_CONSTANTS };
 
 export type RadioProps = Props;
 export type RadioGroupProps = GroupProps;
-export { Radio };
 export default RadioGroup;
